refactor: extract COI warning banner into separate component (#86)
Extract COI service worker warning banner from App.vue into
CoiWarningBanner.vue. App.vue reduced from 152 to 41 lines.
All banner logic, state, and styles are now self-contained.

Closes #86
